requirements.txt
discord.py==2.3.2
python-telegram-bot==21.3
aiohttp==3.9.3
python-dotenv==1.0.1